# [Performance] Prevent unnecessary string concatenation in toHTML.