I've been concatenating strings like a noob. Thanks to profiling, this PR prevents a boatload of allocations, bringing my test parse time down from 17s to under 5s, when generating HTML output. As a test, I commented out all output generation, which saved about half a second. So the parsing itself takes about 4.5 seconds and generating and writing the 24 MB HTML file half a second (on an SSD).

The bad news: generating text output (34 MB in size, via `parseTree.toString.toFile`) takes 28 seconds. So the problem is even bigger there. Arguably, this isn't that important as these massive text files are rather useless anyway, but it is a nice illustration of how a few string concatenations in the wrong place can bring down performance.